backend/requirements.txt
django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.2.2
google-genai==1.66.0
